+++ CHANGELOG.md
@@ -9,6 +9,7 @@ changed functionality, fixed bugs).
 - Windows compatibility fix when using mapped network storage.
 - Windows compatibility fix when using different assets with the same path but on different drives.
 - Allow setting the Shaman JWT authentication token in the `SHAMAN_JWT_TOKEN` environment variable.
+- Blender 2.81 compatibility fix.
 
 
 ## Version 1.1.1 (2019-04-18)